requirements.txt
psycopg2-binary==2.9.6
Flask==2.3.2
python-telegram-bot==13.7